+++ tensorflow/core/platform/s3/s3_file_system.cc
@@ -36,7 +36,6 @@ limitations under the License.
 #include <aws/s3/model/HeadBucketRequest.h>
 #include <aws/s3/model/HeadObjectRequest.h>
 #include <aws/s3/model/ListObjectsRequest.h>
-// #include <aws/s3/model/PutObjectRequest.h>
 
 #include <cstdlib>
 
@@ -535,33 +534,7 @@ Status S3FileSystem::Stat(const string& fname, FileStatistics* stats) {
 
 Status S3FileSystem::GetMatchingPaths(const string& pattern,
                                       std::vector<string>* results) {
-  log_system_->LogMessage(Aws::Utils::Logging::LogLevel::Info, "Get matching paths " + pattern);
-  results->clear();
-  string bucket, object;
-  string fixed_prefix = pattern.substr(0, pattern.find_first_of("*?[\\"));
-  TF_RETURN_IF_ERROR(ParseS3Path(fixed_prefix, false, &bucket, &object));
-
-  Aws::S3::Model::ListObjectsRequest listObjectsRequest;  
-  listObjectsRequest.WithBucket(bucket.c_str())
-      .WithPrefix(object.c_str());
-  listObjectsRequest.SetResponseStreamFactory(
-      []() { return Aws::New<Aws::StringStream>(kS3FileSystemAllocationTag); });
-  auto listObjectsOutcome =
-      this->GetS3Client()->ListObjects(listObjectsRequest);
-  if (listObjectsOutcome.IsSuccess()) {
-    auto contents = listObjectsOutcome.GetResult().GetContents();
-      // Match all obtained files to the input pattern.
-    for (const auto& f : contents) {
-      if (Env::Default()->MatchPath(string(f.GetKey().c_str()), pattern)) {
-        results->push_back(string(f.GetKey().c_str()));
-      }
-    }
-  } else {
-    return errors::Unknown("ListObjects failed with the error",
-                            listObjectsOutcome.GetError().GetExceptionName(), ": ",
-                            listObjectsOutcome.GetError().GetMessage());
-  }
-  return Status::OK();
+  return internal::GetMatchingPaths(this, Env::Default(), pattern, results);
 }
 
 Status S3FileSystem::DeleteFile(const string& fname) {
